Attempt to draw a triangle with modern opengl

+++ Timer.cpp
@@ -26,7 +26,7 @@ double Timer::elapsedMilliseconds()
         endTime = m_EndTime;
     }
     
-    return std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
+    return static_cast<double>(std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count());
 }
 
 double Timer::elapsedSeconds()
